refactor: extract app shell model helpers
